POC: Lazy installation of plugins
When installing plugins, they are not downloaded right away but only
configured and marked as "ready to be installed".
Then, when the user runs a plugin, it gets installed if not installed
already.

After testing, I don't think this UX is better; it is actually probably
more annoying to want to use a plugin and suddendly have to wait for it
to get downloaded.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>