Don't refresh the plugin inventory DB too often
NOTE: with this feature, we may not need a period DB refresh for the
essentials plugins.  We could instead have the essentials plugin NOT
force using the cache, and then we would get a DB refresh no more than
the interval of this feature.  So, if a user didn't run the CLI for a
month, the very first command would check the essentials plugins and
would automatically refresh the DB. We just have to see if we can make
the interval long enough that the DB refresh on some random CLI command
will not be problematic

Signed-off-by: Marc Khouzam <marc.khouzam@broadcom.com>