Passthrough cache towards an OCI registry
My idea is to set this up in the CLI-Svc and point it to the OCI
registry that is setup in TPSM.

This passthrough cache would allow a well-know address to be used
from every end-user machine and also not require direct access to the
real OCI registry.

What is missing is to restrict the passthrough to the tanzu_cli/plugins
directory of the full OCI registry.

To use this test oci passthrough:

make start-passthrough-cache
tanzu plugin source update default \
          -u localhost:5555/tanzu_cli/plugins/plugin-inventory:latest

Signed-off-by: Marc Khouzam <marc.khouzam@broadcom.com>